requirements.txt
httpx
python-dotenv
pytest
python-telegram-bot
aiosmtplib
faiss-cpu
feedparser